Feature/test refactoring

+++ src/main/java/com/example/Library/service/BookCopyServiceImpl.java
@@ -5,11 +5,13 @@ import com.example.Library.exception.NotFoundException;
 import com.example.Library.model.dto.BookCopyDto;
 import com.example.Library.model.entity.Book;
 import com.example.Library.model.entity.BookCopy;
+import com.example.Library.model.entity.BookRental;
 import com.example.Library.model.mapper.BookCopyMapper;
 import com.example.Library.model.mapper.BookMapper;
 import com.example.Library.model.mapper.UserMapper;
 import com.example.Library.repository.BookCopyRepository;
 import com.example.Library.repository.BookRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -65,7 +67,10 @@ public class BookCopyServiceImpl implements BookCopyService {
     @Override
     public void delete(Long bookCopyId) {
         BookCopy bookCopy = findBookCopyModel(bookCopyId);
-        bookCopyRepository.delete(bookCopy);
+        if (!bookCopy.isRented())
+            bookCopyRepository.delete(bookCopy);
+        else
+            throw new BadRequestException("Is rented");
     }
 
     @Override
